kjgags

+++ springsecuritybyeazybytes_6/springsecuritybyeazybytes_6/src/main/java/com/springsecuritybyeazybytes/repository/CustomerRepository.java
@@ -3,18 +3,14 @@ package com.springsecuritybyeazybytes.repository;
 
 import com.springsecuritybyeazybytes.model.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
-
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
-       List<Customer> findByEmail(String email);
-
-
-
-
-
+    List<Customer> findByEmail(String email);
+    
 }
